+++ HISTORY.md
@@ -3,8 +3,8 @@
 ## 0.8.0 (Unreleased)
 * Lazily initialize and reuse SQLite connection objects
 * Fix `AttributeError` when using a response cached with an older version of `attrs`
-* Add `fast_save` option for SQLite backend (`PRAGMA` setting to improve write performance, with
-  some tradeoffs)
+* Fix concurrent usage of `SQLiteCache.bulk_commit()`
+* Add `fast_save` option for `SQLiteCache` (`PRAGMA` setting to improve write performance, with some tradeoffs)
 
 ## 0.7.3 (2022-07-31)
 * Remove upper version constraint for `attrs` dependency